# ddd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/demo/api/v1/demo": {
            "post": {
                "description": "数据库增接口demo的描述内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "数据库增接口demo",
                "operationId": "InsertDemo",
                "parameters": [
                    {
                        "description": "AddDemoReq",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.AddDemoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.JSONResult"
                        }
                    }
                }
            }
        },
        "/demo/api/v1/demos": {
            "get": {
                "description": "数据库查接口demo的描述内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "数据库查接口demo",
                "operationId": "GetDemos",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "demo_name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetDemosResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddDemoReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "demo_name"
                }
            }
        },
        "dto.Demo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "dto.GetDemosResp": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "demos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Demo"
                    }
                }
            }
        },
        "dto.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}